Asynchronous plotting for Theano (#124)
Added support for asynchronous plotting for Theano. The main caveat is that Linux machines need to use multiprocessing.Process and Mac OS X machines need to use threading.Thread. If Linux machines use Threads, the program slows to a crawl (like as a result of Python's Global Interpreter Lock GIL); conversely, if Mac machines use Process, the glfw will throw a segmentation fault and fail to draw the window.