Add errors to the absolute and incremental checks
Error codes have been added to both the absolute and incremental checks
in file check_flake8.sh.
An option was added to the former mentioned script to run the
incremental verification locally so developers can use it to look for
errors before pushing the code for verification in Travis.
The order of select and ignore in flake8 was swapped to allow filtering
specific error codes, specifically, in for naming conventions (N) and
docstrings (D).
The package flake8-quotes was replaced for flake8-docstrings, since both
can check that docstrings use triple double quotes. Also the package
pep8-naming was added to check style rules for naming conventions.
Finally, the test files for PEP8 were removed.