Change string of errors into array in check_flake8
The string of errors in check_flake8 was replaced by an array to improve
readability. Also, the variable "errors" was replaced for two variables
that correspond for the incremental and absolute tests.
The pip packages required for docstring and naming conventions were
migrated to environment.yml, and the script for pylint was removed,
since it was considered for indentation but it has conflicts with yapf.