Fix style errors and add ignored errors
The variable errors_indentation had a typo and was changed to
errors_incremental. The style errors include wrapping all lines to 80
characters and indentation of two spaces.
The ignored errors are arrays for error codes that have to be ignored
for both the incremental and absolute checks. Currently they're empty
but they are defined for documentation purposes.