Fix tensorboard creates another TF device issue
Class TensorBoardOutput used to create tf.Session in its constructor.
But the session seems to launch a TF device. Since the only use of the
session is to dump histogram, and usually there is a session created as
default when dump histograms into tensorboard. So I deleted it in the
constructor and moved it into method _dump_histogram.

There is also a circular import in TensorBoardOutput, deleted it.