Add OU exploration strategy to TensorFlow
Ornstein Uhlenbeck exploration strategy comes from the
Ornstein-Uhlenbeck process, which is a stationary process that describes
the velocity of a Brownian particle under the influence of friction. The
OU strategy is often used in DDPG algorithm because in continuous
control task it is better to have temporally correlated exploration to
get smoother transitions. And OU process is relatively smooth in time.