Add replay buffer to TensorFlow
Replay buffer is an important technique in reinforcement learning. It
stores transitions in a memory buffer of fixed size. When the buffer is
full, oldest memory will be discarded. At each step, a batch of memories
will be sampled from the buffer to update the agent's parameters. In a
word, replay buffer breaks temporal correlations and thus benefits RL
algorithms.

DDPG uses replay buffer to update parameters in policy network and q
function network.

This commit implements a ReplayBuffer class to TensorFlow. In each step,
it adds a transition of (observation, reward, terminal, action,
next_observation) into the memory buffer. And returns a dict of a batch
of transitions when agent asks for a sampled transitions.