Implement actor-critic for DDPG
DDPG uses actor-critic method to optimize the policy and reward
prediction. The actor-critic method a kind of TD method. It consists of
an actor, which is a policy network used for action selection, and a
critic, which is a q function network used for action value estimation.
The critic genetates TD error to indicate the optimization of actor and
critic.

In DDPG, actor accepts observation as input, uses MLP to fit the mapping
from observation to action, then outputs the predicted action. The
critic accpets observation and action as the input, output the q value.

Here, the policy network is named as ContinuousMLPPolicy. The q function
network is names as ContinuousMLPQFunction.