Override plot enable from function run_experiment
The plot parameter cannot be directly modified since it's embedded in an
anonymous function that is serialized for cloud processing (bytecode
cannot be modified), so a static method was used instead to override the
plot enabling of the algorithm constructor with the plot enabling of the
function run_experiment.
Also the script plotter.py was modified to convert it into a class like
the one in the TF branch.