Call set_enable for both Theano and TF branches
Since the method set_enable exists for both the Theano and the TF
branch, but there's no clear way to distinguish the branch used in
function run_experiment, the method must be called for both branches.
Also, the enable variable was missing from the Plotter class in the TF
branch.