Change set_enable to disable and remove status
For the Plotter class in the Theano and TF branches, the static method
set_enable was changed to disable and the method status was removed. The
guards for Plotter.enable were simplified to only enabled the functions
when the flag is true.
The plot variable in the algorithm constructors is equal to itself OR
Plotter.enable to consider the scenario where plot=False in the
algorithm and plot=True in run_experiment, considering plot should be
overriden with the value defined in run_experiment.