Add DDPG to TensorFlow (#66)
* Add OU exploration strategy to TensorFlow

Ornstein Uhlenbeck exploration strategy comes from the
Ornstein-Uhlenbeck process, which is a stationary process that describes
the velocity of a Brownian particle under the influence of friction. The
OU strategy is often used in DDPG algorithm because in continuous
control task it is better to have temporally correlated exploration to
get smoother transitions. And OU process is relatively smooth in time.

* Add replay buffer to TensorFlow

Replay buffer is an important technique in reinforcement learning. It
stores transitions in a memory buffer of fixed size. When the buffer is
full, oldest memory will be discarded. At each step, a batch of memories
will be sampled from the buffer to update the agent's parameters. In a
word, replay buffer breaks temporal correlations and thus benefits RL
algorithms.

DDPG uses replay buffer to update parameters in policy network and q
function network.

This commit implements a ReplayBuffer class to TensorFlow. In each step,
it adds a transition of (observation, reward, terminal, action,
next_observation) into the memory buffer. And returns a dict of a batch
of transitions when agent asks for a sampled transitions.

* Implement actor-critic for DDPG

DDPG uses actor-critic method to optimize the policy and reward
prediction. The actor-critic method a kind of TD method. It consists of
an actor, which is a policy network used for action selection, and a
critic, which is a q function network used for action value estimation.
The critic genetates TD error to indicate the optimization of actor and
critic.

In DDPG, actor accepts observation as input, uses MLP to fit the mapping
from observation to action, then outputs the predicted action. The
critic accpets observation and action as the input, output the q value.

Here, the policy network is named as ContinuousMLPPolicy. The q function
network is names as ContinuousMLPQFunction.

* Add DDPG to TensorFlow

Deep Deterministic Polict Gradients is an off-policy reinforcement
learning algorithm for contiunous control tasks. It uses the replay
buffer and two target networks, corresponding to policy and q function
network, to stablize the training process.

This DDPG algorithm uses the primitives in garage/tf to set up the
algorithm.

* Add launcher file for DDPG to TensorFlow

* Add miscellaneous change to DDPG

 - Fix bugs in DDPG. The last DDPG incorrectly calculates action_loss.
 - Set output_nonlinearity param to tf.nn.tanh in ddpg_pendulum.py.
   Without this, the training process will be negatively influenced.
 - Delete mistakenly pushed files.

* Reorder imports and add docstring

 - Add missed args in docstring
 - Reorder imports in alphabetical order

* Add tf async plotter to DDPG

 - Add tf async plotter to DDPG
 - Formatting other files

* Improve training process of DDPG

Add transitions to DDPG when terminal is True. This proves to improve
the average return significantly.

* Reformatting ddpg.py to conform with flake8

 - Add multiple docstring to ddpg.py.
 - Change get_target_ops to module methods since it is rather a static
   method than a class method.

* Break lines in q_functions/__init__.py

Break lines in garage/tf/q_functions/__init__.py because line is too
long.

* Add multiple docstrings to multiple files

There are several files has missing docstrings. Add them to pass flask8
checks.

* Add miscellaneous changes

 - Fix errors in docstring.
 - Reorder imports.
 - Add requirement third party lib baselines into environment.yml.

* Add benchmark test into tests

Add a benchmark script which can run regression tests between baselines
and garage algorithms.

For different algorithms, imports, params, method run_garage() and
run_baselines() need to change.

* Add requirement lib for mpi4py

 - Mpi4py needs libopenmpi-dev be pre-installed in Linux.

* Change tf_benchmark_ddpg plotter

 - Change plotter in tf_benchmark_ddpg.py from one plot per trail to one
   plot per task, one curve per trail.

* Reformat code with YAPF

* Add another dep of mpi4py

Add openmpi-bin to deps so that the check_imports passes.

* Add title into benchmark plotter

* Fix syntax error in q_functions/__init__.py

* Add miscellaneous change

 - Use flat_dim to get dimenstion of space.
 - Change error comment in the class docstring of DDPG.
 - Use tf.set_random_seed to get same sequence of values.
 - Delete trainable=True in the build_type().

* Fix D413 in multiple files

There should be a blank line after last section of docstrings.

* Add miscellaneous change

 - Use bounds in garage.envs.util to get action spcace bound.
 - Set built_net to private.
 - Namespace log in DDPG to a relevant component.
 - Move benchmark.png to the parent directory, more convenient.
 - Set default params of DDPG to the one gets best performance.
 - Refactor critic_optimizer name in DDPG.