Make joblib use the "forkserver" start method
As indicated in joblib documentation, to configure the start method as
forkserver, the JOBLIB_START_METHOD environment variable has to be set
as "forkserver". In garage, this is now done at the beginning of the
run_experiment function, so the child process that runs joblib has the
start method configured as forkserver.
An assert is added to the child process that runs the experiment to make
sure the variable is correctly set.