Cleanly terminate worker processes on an interrupt
When calling function run_experiment and the parameter n_parallel is
bigger than zero, worker processes are created to do parallel sampling
during the training. However, when there's a keyboard interrupt, there
are certain occasions when processes are not completely killed after the
execution has finished.
By checking the PID of these processes using the command "ps -fu | grep
run_experiment", we noticed they had the status code "Sl", where S means
interruptible sleep, so the processes are waiting for a signal to wake
up and continue their execution.
When executing the command "kill -SIGINT <pid>" with the PID of the
sleeping process, we got the python trace back and found that the
processes were sleeping trying to acquire the lock of the inqueue inside
the pool of python multiprocessing.
By default, when a keyboard interruption occurs all processes get
notified, but the notifications are not sent in any order. This could
produce an issue when a process is interrupted and finished without
releasing the lock that other process are waiting for as it's in this
case.
To avoid this problem, the parent of the worker processes catches the
BaseException (consider other cases besides the KeyboardInterrupt) and
terminates makes sure to terminate the pool of worker processes to avoid
the dead locks.
Also, the start method of joblib was configured as forkserver. As
indicated in joblib documentation, the JOBLIB_START_METHOD environment
variable has to be set as "forkserver". In garage, this is now done at
the beginning of the run_experiment function, so the child process that
runs joblib has the start method configured as forkserver.
An assert is added to the child process that runs the experiment to make
sure the variable is correctly set.
Other changes were required to enforce the PEP8 style in legacy code.