Missing changes to cleanly exit worker processes
The worker processes pool is only created when n_parallel is bigger than
1, so n_parallel=1 was creating null exceptions when terminate was
called. This is solved by adding a guard to check if the pool was
instantiated.
Another fix that is required is to rename back the member g in
singleton_pool to G. This was done in a previous change to enforce PEP8
style of having non-capital variables/parameters, but the member has to
keep the same name as defined in singleton_pool.