Remove docstring references to run_experiment
The references of run_experiment in the docstrings of plotter were
removed. Also the plot parameter of the algorithms is not modified by
the external parameter Plotter.enable anymore, since the logic of the
algorithm has to remain independent from that of the Plotter.
Also, the guard patter was used in the methods of the Plotter class.