Reraise exceptions caught in run_experiment.py
We catch BaseException in run_experiment so that we can cleanly
terminate all of the worker processes when an interrupt occurs. If we
don't re-raise the exception after cleaning up, all exceptions which
end the program are silently supressed.

This change re-raises the caught exception after clean-up so that it
propagates to the user.