Add a GitHandler to switch to a target commit
The new class GitHandler is incorporated along with the function
run_experiment to allow the users to run experiments at a desired commit
based on a branch name (remote or local), tag name or SHA. The parameter
passed to run_experiment is a key value, for example:
    branch: feature_branch
    branch: origin/master
    tag: baseline_052318
    sha: 4d93a274bcd2de0
Besides this specific feature, GitHandler provides methods to interact
with simple git operations for future features.
Also, a test file is provided to make sure the implementation of
GitHandler is working correctly.