Use tf.variable_scope() in FirstOrderOptimizer (#154)
TF optimizers add gradient annotation variables (e.g. foo/Adam:0) to the
variable scope of the TF ops associated with the gradient. This creates
a problem for Parameterized classes, because constructing the
Parameterized class in a new process will *not* construct the
optimizer as in the main process, meaning that these annotation
parameters will be missing. This causes the class to fail to serialize
across processes.

Enclosing FirstOrderOptimizer's TF op creation in tf.variable_scope()
quarantines the new annotation parameters in a new variable scope,
away from subgraphs of Parameterized primitives.