Remove /progress from Tensorboard output paths
There is an argument in scripts/run_experiment that sets
log_dir/progress as the tensorboard output dir. Delete it and set the
tensorboard dir same as log_dir removes `/progress` in the end of every
run.

Also, mkdir_p in TensorboradOutput is useless and it creates an empty
dir every run when rencord_tensor is not called. Delete it.