Add default values for GCS bucket and path
The default values are used by run_experiment, and to indicate this
function to save the logs in GCS, a boolean variable is set instead.
Missing documentation was added, and the strategy to remove the path
from the files uploaded in the bucket was changed to completely remove
all directories from the file path and only use the path passed as
parameter to function upload_to_gcs.