Add PPO to garage/tf
Add Proximal Policy Optimization algorithm into garage TensorFlow tree.
PPO can be inherited from NPO. The difference among PPO, NPO and TRPO is
the constraint of the surrogate objective. This commit rewrites the
current NPO in garage tf so that it can run in GPU.

The policy, baseline and optimizer that PPO needs are already
implemented in the primitives of garage.

The process_samples function in garage.samplers.BaseSampler only
produces dense trajectories for RNN implementation. But it is faster if
we always produce dense trajs and flatten them in GPU. This commit adds
a process_samples function in garage.tf.samplers.BatchSampler and make
it the default sampler for TF tree. The VectorizedSampler inherites it.

Also rewrite the TRPO in garage/tf since the NPO implementation has
changed.

Add flatten functions into garage/tf/misc/tensor_utils.py.

Summary of what this commit does:
 - Add PPO implementation to garage tf tree.
 - Add example file of PPO into examples/tf/ppo_pendulum.py
 - Add benchmark tests of PPO and TRPO.
 - Add unit tests for garage.tf.misc.tensor_utils.
 - Add unit tests for garage.tf.algos.