Move Theano-dependent code in the misc package
The module special.py is used by both the Theano and TensorFlow
packages, but it contained some Theano dependent functions. Only those
functions were moved into garage.theano to keep the rest of the
functions for both distributions in the new module tensor_utils.py.
The same thing was done for ext.py.
The rest of the modules in garage.misc are independent of Theano.