Update setup_linux.sh based on Dockerfile.ci (#258)
The dependencies used in docker to build garage in Travis CI have been
added to the script to setup garage in Linux.
Also, with the help of Argbash the following arguments have been added:
- --tf-gpu: installs the GPU flavor of TensorFlow instead of the vanilla
flavor.
- --set-envvar: it writes the file ~/.bashrc to set the environment
variables required by garage. Otherwise, the user is notified of the
required variables to modify.
- mjkey-path: the path to the MuJoCo license key.

The current script has been tested in a fresh installation of Ubuntu,
running the files under the examples folder (both Theano and TensorFlow)
right after to make sure this setup file is working.