Move DmControl{Env, Viewer} to a separate package (#269)
This PR moves DmControlEnv and DmControlViewer to their own package,
so that they are not automatically imported any time garage.env is
imported (which, for practical purposes, is always). Previously, this
meant that effectively all launchers import dm_control and pygame,
slowing the launch process, increasing memory footprint, and making
launchers more prone to crashing.