Fix script to run in OS X 10.12
The command chmod and the conda installer are no longer called as root
user since it's not required and it creates user access issues in the
conda folder with root.
Also, pybo2d is compiled from source and gcc downgraded to version 7 in
order to fix the compatibility issue. In this way, garage works right
after the setup script finishes successfully.