Update setup script for OS X (#279)
The script has been updated based on setup_linux.sh, using homebrew to
install Linux packages and updating the requirements for mujoco-py, gym
and baselines according to their documentation for OS X.
Since mujoco-py is installed within the setup scripts for Linux and OS
X, the script to setup mujoco has been removed.
Also, the tensorflow package was added to environment.yml so it can be
installed out of the box without the scripts. The feature to install the
GPU flavor of TensorFlow may be removed from the Linux script once
everything can be installed using environment.yml only, or another
single list of dependencies.
Also, bash primitives replaced readlink, using directories instead
of suffixes with mktemp and fixing some output messages.
Some of the fixes for OS X were applied to the Linux scrip as well.
Finally, the default for set-envvar and the correct replacement of
string in print error and warning functions have been added to the
script to setup linux.