Test to interrupt experiment at different stages
The class BatchPolopt has been overridden as BatchPoloptCallback to
notify the test of the different stages in the experiment life cycle so
it can be interrupted with SIGINT.
The test makes sure that the children processes produce are zero after
the SIGINT is sent, or it throws an assertion error with those processes
that didn't die.
Also the context manager MasksSignals has been created to handle the
masking of SIGINT in parallel_sampler.