Shutdown child processes during interruption
All plotters are appended to a static list, so they can be easily
reachable from run_experiment to call shutdown on them.
Also, terminate was replace by close to shutdown the parallel sampler,
since terminate calls join and may block the shutdown in run_experiment.
In order to check that all processes died after a time out, a loop to
poll for alive processes was implemented.
A warning message for processes that remain after shutdown is printed so
users of garage can reopen the corresponding issue.