Remove semaphore tracker and close Manager
A process known as semaphore tracker is spawned from the run_experiment
process, but we cannot stop this process as it ignores SIGINT and
SIGTERM and we haven't access to it. Therefore, it's removed from the
list of children to wait for in both run_experiment and in test_sigint.
Another process that is spawned from run_experiment is the Manager,
which owns multiprocessing objects such as RLocks and counters used
during the run_collect method in the StatefulPool class. This process
was also making the warning message appear and the test fail.
However, the manager has shutdown method to terminate the process, so we
can verify the termination of the process.