Move GatherEnvs from mujoco_rgb_classes
In TestEnvs, some MuJoCo environments are tested and rendered, asserting
that they're producing a viewer object. Since the rendering for Gather
environments is bypassed until is fixed (see issue 323 in garage), the
previous mentioned assertion fails.
For that reason the Gather environments were moved from the
mujoco_rgb_classes to the mujoco_classes in TestEnvs.