Import mujoco_py in GarageEnv only if necessary (#340)
Currently the viewer window for MuJoCo environments is not closed
correctly, leading to out of memory issues. This bug is on the side of
the gym environments, but until gym provides a fix, we're closing the
viewer window directly.
However, by importing the mujoco_py package into the environment base of
garage, this enforces the use of MuJoCo for all environments, which is
not necessary for non-MuJoCo environments.
That's why the import for the mjviewer was moved into the exception
where the closing of the mjviewer is performed.