Fix TensorFlow warning in NPO about IndexedSlices (#375)
NPO uses a boolean valid_var tensor to select a slice from a variety
of tensors related to action, policy and advantages. Previously we
used tf.boolean_mask() to do the selection, which internally calls
tf.gather() that emits tf.IndexedSlices. When calculating BP,
tf.IndexedSlices will be casted into a dense tf.Tensor with a
"Converting sparse IndexedSlices to a dense Tensor of unknown shape"
warning. This fix replaces tf.boolean_mask() with
tf.dynamic_partition(), which don't have such problem.

This difference of behavior is not documented in Tensorflow. It is
not clear if tf.dynamic_partition() will cause performance
regression.

This fixes #289.