Conditionally import lasagne or tf to set the seed
When launching experiments with run_experiment the function set_seed is
in charge of setting the random seed across the different packages used
in garage. However, we should keep from import theano when only using
tensorflow and vice versa.
To detect if theano and/or tensorflow are being used, we check if they
appear in the list of packages already imported.