Add dockerfiles for garage
This commit introduces commands to build and run docker containers with
garage. Currently there's three types of containers:
 - ci: it uses the same docker containers as headless, but the 
   corresponding build and run commands are configured to be used with
   Travis CI. 
 - headless: it doesn't provide visualization of the environments used
   in garage.
 - nvidia: it provides visualization and support for NVIDIA GPUs.

The commands to build and run these containers are defined in the 
Makefile at the root of the repository.
The file "Dockerfile.ci" was split in three files: Dockerfile.headless,
which contains specific instructions to build the headless and ci
images, Dockerfile.nvidia, which has instructions to build the nvidia
image, and Dockerfile, which has base instructions for the two previous
images. Also, docker-compose is now used to build the images.
For further details, please read docker/README.md.
The configuration in Travis to run the docker container was also
modified to now run the ci container using the Makefile.