Fix caching and CodeCov in garage-ci
The docker image for garage-ci was having the following issues:
- The codecov environment variables were not expanding correctly when
calling docker-run though a makefile in Travis.
- Caching was not working
First, the ci image got its own compose file docker-compose-ci to simply
things. To solve the codecov issue, the (bash<curl) operation was moved
from the makefile to the travis file, since the operation was not 
expanding properly when called from the makefile.
To pass the environment variables of codecov generated in travis, the 
variable ADD_ARGS was added in the makefile to pass these or any other
environment variables required in the future. This was documented in the 
README.md.
For caching, the wrong tag was used, since ":latest" was getting
appended to the name tag, giving an invalid value for cache image name.
Since each image type has its own compose file now, the tags for caching
are now hard-coded to their corresponding image types.
Also, only the presence of the mjkey.txt is needed to install mujoco-py, so an
empty file named mjkey.txt is created in the Dockerfile instead of passing the
real key.
Finally, the file Dockerfile was renamed to Dockerfile.base.