Use only CodeCov for coverage tracking
In #316 we introduced several code coverage tracking providers to the
CI pipeline, and resolved to remove all but one of them after a trial
period.

CodeCov has proven to be the clear winner for our project and workflow,
so I'm removing all of the other reporters. This will hopefully
simplify the CI pipeline a little bit.