Exit with a non-zero status in check_tests script
Both scripts check_test and check_cronjob_test are used when running
Travis CI, and it was observed that the CI was still passing if any of
the commands in the scripts failed.
The argument -e is now set in both scripts to exit the scripts with non-
zero status if case any of their commands fail.