Fix some logging issues in the new local runner (#562)
This commit makes run_experiment() back in examples and fixes some other logging issues.

Changlog

* Use run_experiment() in examples to set up the logger to use tensorboard.
* Remove logging of histogram of actions.
* Fix ProcessExecTime not logged issue in on-policy sampler.

Note
Logging of action histogram is removed because tensorflow only supports logging 
histogram of a tensor of fixed shape, but the number of actions is dynamic in each epoch.

Previously we logged the histogram of actions by truncating samples to the minimum
number of trajectories (=batch_size / max_path_length). This is not possible because
the algorithm doesn't know batch_size anymore. The old way of logging actions was
also not accurate, because it logged after the max_path_length zero padding.