Make the experiment wrapper a module
This makes it easier to use garage as a library. It makes the
experiment wrapper script available inside the garage package path.

To avoid name collisions, this PR also renames the experiment wrapper
from `run_experiment.py` to `experiment_wrapper.py`.

Some hacky changes to garage.experiment are necessary to make this
work, but we will be deleting that module (and this wrapper) ASAP.