Refactor CEM under the new interface (#593)
The old CEM implementation in rllab interacts directly with singleton_pool,
with each worker collecting rollout for a sampled policy. 
This doesn't use Sampler, which is not desired, and is not compatible with
the new train_once() interface. There was also some issues with tf initialization
in the worker that made CEM not available. See issue #460 .

This commit fixes CEM by doing rollouts for one sampled policy at each epoch
cycle and updating policy distribution at the end of each epoch.