Use scipy.stats distribution types in the logger
Replaces tensorflow_probability types used in TensorBoardOutput with
equivalent types from scipy.stats. Also adds a simple mechanism for
logging histograms of raw samples called `garage.logger.Histogram`.

The origin of this PR is a critical performance implication of using
tensorflow_probability types to represent distributions. In practice,
algorithms will need to create a new tfp.Distribution object every time
they log a distribution value. The logger cannot evaluate most ops in
the algorithm's TF graph, because they require a feed_dict to evalaute.

To get around this, most people would just construct a new Distribution
op with their concrete data:
```python
with tf.name_scope('logging'):
    dist = tfp.distributions.MultivariateNormalDiag(
        loc=means, scale_diag=stds)

tabular.record('Normal', dist)
```

Unfortunately, this modified the TF graph which is slow. Repeated
modifications to the graph appear make the runtime of `tf.Session.run`
scale quadratically in the number of modifications. This slows
algorithms to a crawl in a few hundred iterations.

I expect this situation might be fixed in TF 2.0, but for now it is
safest to remove this feature from the logger and use scipy.stats,
which provides just as much expressiveness.