Add tf/DQN with dueling support (#582)
DQN implementation with garage.Model.

This is the first algorithm for pixel environments. This PR
adds the algorithm as well as the models, primitives and
environment wrappers required for training in pixel environments.

* Models
  * MLPDuelingModel
  * CNNModelWithMaxPooling
* Primitives
  * QFunction2 (base class, without parameterized)
  * DiscreteCNNQFunction
* Wrappers
  * AtariEnvWrapper (needed when using env wrappers from baselines)

The eviction policy of replay buffer used to be random. To
make experiments determinisitic, it is changed to
First In First Out (FIFO). It was proven to be necessary in
order to achieve better result in complex environment for DQN.

Added corresponding properties in QFunction to make reference
easier, e.g. q_vals.

Added clone method in QFunction to enable copying configuration,
not including the parameters. Since all the necessary operations
will be done in object construction, we can simply create a new
object with a different name (because we want the new object
to have a different name for variable scope).