Fixup between LocalTFRunner and Ray Sampler PRs (#830)
* Fixup between LocalTFRunner and Ray Sampler PRs

I merged the Ray Sampler without updating because of LocalTFRunner
changes. This fixes the broken example (and does some minor cleanup).

* Fix RaySampler test failure

It appears that there's a behavioraly difference between ray 0.7.3.
Specifically, `LocalModeObjectID`s passed into remote methods aren't
implicitly requested in 0.7.3, but are in 0.7.2. It's not clear what the
intended behavior here is, so we're going to avoid it altogether for
now.
This version still pickles the environment and agent, then transmits
them. In the environment's case this is just and optimization to avoid
repeated pickling, but in the agent's case this is needed, since the TF
session needs to be set up when it unpickles.

This change might have a small performance impact when starting the
sampler, but doesn't affect the steady-state performance, so it should
be fine to merge.