Add a new ReplayBuffer, "PathBuffer" (#838)
This is a replay buffer implementation that differs from the existing
replay buffer implementation in that it can efficiently store variable
length paths. Ultimately, it should replace the existing replay buffer
and allow us to fix #837.

It may also be useful for as yet unimplemented off algorithms which want
to process whole paths.