Fix CI for branch builds; Update flake8-docstrings (#889)
* Fix CI for branch builds

Currently check_precommit.sh fails for release branches because it
checks all commits that are different between master and branch.
However, not all commits in master follow the commit message
guidelines since those messages are made through github squash and merge
interface. Also, there's no point to checking commits that are already
in master.

This commit modifies check_precommit.sh to pick only commits in the
branch that are not in master, by taking the commits in branch that were
made after the last common (merge-base) commit of master and branch.

Test run in a branch cut from release-2019.02:
https://github.com/rlworkgroup/garage/commit/5d4ce2360db2000c83a397ac0bd9a9445b8af007
https://travis-ci.com/rlworkgroup/garage/builds/128575672

* Update flake8-docstrings and restrict pydocstyle

flake8-docstrings==1.3.0 breaks with pydocstyle>=4.0.0, but this was
fixed in 1.3.1 (https://gitlab.com/pycqa/flake8-docstrings/issues/36).
 We fixed this in master by restricting pydocstyle to
<4.0.0. But the issue is also present in previous release branches,
breaking the CI.

This commit updates flake8-docstrings to current functioning version
1.4.0 and restricts pydocstyle to bring it up to date (4.0.1) while also
restricting it to <4.1 to prevent future issues.