Add separate Travis job for pre-commit checks
This PR adds a separate CI job for pre-commit checks. This makes it
easier to distinguish at a glance pre-commit failures from test
failures. It also allows us to disable this job for non-PR builds.

The downside of this PR is that it adds an additionao build job and
~5 minutes of overhead to the PR build process, because it must pull
the Docker cache. For builds which update the Docker container, this
will add an additional ~10 minutes to rebuild the container. This can
be addressed in a later PR by adding a separate build stage to build
and cache the container.