Constructing unit tests for sac_agent_losses
Currently, there is some bug that is causing the policy loss to not
greatly decrease with time. This is causing the learning in later
iters to be much slower than it actually should be.

To examine this, I am porting over the loss tests from tf_agents/sac
to this code base, with the hopes that it will reveal any issues
in the loss/objective functions in garage/sac