Add TrajectoryBatch data type
Recent experience has shown that many bugs and API ambiguities could
be avoided if some of the abstract data types passed around by
components of garage were well-defined. This PR attempts to define one
such data type, TrajectoryBatch, which represents trajectory data in
the form frequently used by on-policy RL algorithms.

An advantage of this approach is that documenting the inputs and
outputs of various parts of the data processing pipeline becomes much
simpler. In the future, we can also add validation checks and other
helpers to these classes to make data type creation and conversion
more predictable and safer.